+++ tcc/routes/web.php
@@ -22,7 +22,7 @@ Route::post('/verirficacodigo', 'LoginController@verirficacodigopost')->name('ve
 
 Route::group(['prefix' => 'aplicacao'], function () {
 
-    Route::get('/home', 'LoginController@home')->name('home')->middleware('checksession');
+    Route::get('/home', 'homeController@home')->name('home')->middleware('checksession');
 
 });
